// src/types/r3eTypes.ts
export enum PitStopStatus {
	/** No mandatory pitstops */
	Unavailable = -1,

	/** Mandatory pitstop for two tyres not served yet */
	UnservedTwoTyres = 0,

	/** Mandatory pitstop for four tyres not served yet */
	UnservedFourTyres = 1,

	/** Mandatory pitstop served */
	Served = 2
}

export enum EFinishStatus {
	/** N/A */
	Unavailable = -1,

	/** Still on track, not finished */
	None = 0,

	/** Finished session normally */
	Finished = 1,

	/** Did not finish */
	DNF = 2,

	/** Did not qualify */
	DNQ = 3,

	/** Did not start */
	DNS = 4,

	/** Disqualified */
	DQ = 5
}
// ...
export enum ETireType {
	Unavailable = -1,
	Option = 0,
	Prime = 1
}

export enum ETireSubtype {
	Unavailable = -1,
	Primary = 0,
	Alternate = 1,
	Soft = 2,
	Medium = 3,
	Hard = 4
}

export enum EEngineType {
	Combustion = 0,
	Electric = 1,
	Hybrid = 2
}

export enum EPitState {
	None = -1,
	RequestedStop = 1,
	Entered = 2,
	Pitting = 3,
	Exiting = 4
}

export enum EPenaltyType {
	Unavailable = -1,
	DriveThrough = 0,
	StopAndGo = 1,
	Pitstop = 2,
	Time = 3,
	Slowdown = 4,
	Disqualify = 5
}

export enum EOvertakingAid {
	Unavailable = -1,
	NotEngaged = 0,
	Engaged = 1
}

export interface IVector3 {
	X: number;
	Y: number;
	Z: number;
}
// ...
export interface ICutTrackPenalties {
	DriveThrough: number;
	StopAndGo: number;
	PitStop: number;
	TimeDeduction: number;
	SlowDown: number;
}
// ...
export interface ISectors {
	Sector1: number;
	Sector2: number;
	Sector3: number;
}

export interface IDriverInfo {
	Name: string;
	CarNumber: number;
	ClassId: number;
	ModelId: number;
	TeamId: number;
	LiveryId: number;
	ManufacturerId: number;
	UserId: number;
	SlotId: number;
	ClassPerformanceIndex: number;
	EngineType: EEngineType;

	Unused1: number;
	Unused2: number;
}

export interface IRaceSessionMinutes {
	Race1: number;
	Race2: number;
	Race3: number;
}

export interface IRaceSessionLaps {
	Race1: number;
	Race2: number;
	Race3: number;
}

export interface IDriverData {
	DriverInfo: IDriverInfo;
	FinishStatus: EFinishStatus;
	Place: number;
	PlaceClass: number;
	LapDistance: number;
	Position: IVector3;
	TrackSector: number;
	CompletedLaps: number;
	CurrentLapValid: number;
	LapTimeCurrentSelf: number;
	SectorTimeCurrentSelf: ISectors;
	SectorTimePreviousSelf: ISectors;
	SectorTimeBestSelf: ISectors;
	TimeDeltaFront: number;
	TimeDeltaBehind: number;
	PitStopStatus: PitStopStatus;
	InPitlane: number;
	NumPitstops: number;
	Penalties: ICutTrackPenalties;
	CarSpeed: number;
	TireTypeFront: ETireType;
	TireTypeRear: ETireType;
	TireSubtypeFront: ETireSubtype;
	TireSubtypeRear: ETireSubtype;

	BasePenaltyWeight: number;
	AidPenaltyWeight: number;

	DrsState: EOvertakingAid;
	PtpState: EOvertakingAid;
	PenaltyType: EPenaltyType;

	// Based on the PenaltyType you can assume the reason is:

	// DriveThroughPenaltyInvalid = 0,
	// DriveThroughPenaltyCutTrack = 1,
	// DriveThroughPenaltyPitSpeeding = 2,
	// DriveThroughPenaltyFalseStart = 3,
	// DriveThroughPenaltyIgnoredBlue = 4,
	// DriveThroughPenaltyDrivingTooSlow = 5,
	// DriveThroughPenaltyIllegallyPassedBeforeGreen = 6,
	// DriveThroughPenaltyIllegallyPassedBeforeFinish = 7,
	// DriveThroughPenaltyIllegallyPassedBeforePitEntrance = 8,
	// DriveThroughPenaltyIgnoredSlowDown = 9,
	// DriveThroughPenaltyMax = 10

	// StopAndGoPenaltyInvalid = 0,
	// StopAndGoPenaltyCutTrack1st = 1,
	// StopAndGoPenaltyCutTrackMult = 2,
	// StopAndGoPenaltyYellowFlagOvertake = 3,
	// StopAndGoPenaltyMax = 4

	// PitstopPenaltyInvalid = 0,
	// PitstopPenaltyIgnoredPitstopWindow = 1,
	// PitstopPenaltyMax = 2

	// ServableTimePenaltyInvalid = 0,
	// ServableTimePenaltyServedMandatoryPitstopLate = 1,
	// ServableTimePenaltyIgnoredMinimumPitstopDuration = 2,
	// ServableTimePenaltyMax = 3

	// SlowDownPenaltyInvalid = 0,
	// SlowDownPenaltyCutTrack1st = 1,
	// SlowDownPenaltyCutTrackMult = 2,
	// SlowDownPenaltyMax = 3

	// DisqualifyPenaltyInvalid = -1,
	// DisqualifyPenaltyFalseStart = 0,
	// DisqualifyPenaltyPitlaneSpeeding = 1,
	// DisqualifyPenaltyWrongWay = 2,
	// DisqualifyPenaltyEnteringPitsUnderRed = 3,
	// DisqualifyPenaltyExitingPitsUnderRed = 4,
	// DisqualifyPenaltyFailedDriverChange = 5,
	// DisqualifyPenaltyThreeDriveThroughsInLap = 6,
	// DisqualifyPenaltyLappedFieldMultipleTimes = 7,
	// DisqualifyPenaltyIgnoredDriveThroughPenalty = 8,
	// DisqualifyPenaltyIgnoredStopAndGoPenalty = 9,
	// DisqualifyPenaltyIgnoredPitStopPenalty = 10,
	// DisqualifyPenaltyIgnoredTimePenalty = 11,
	// DisqualifyPenaltyExcessiveCutting = 12,
	// DisqualifyPenaltyIgnoredBlueFlag = 13,
	// DisqualifyPenaltyMax = 14
	PenaltyReason: number;

	// -1 unavailable, 0 = ignition off, 1 = ignition on but not running, 2 = ignition on and running
	EngineState: number;

	/** Reserved data */
	Unused1: number;
	Unused2: number;
	Unused3: number;
}